ref(src): Rename task to item in package
