+++ src/main/java/com/cloud/chatbot/Cloud.java
@@ -5,11 +5,11 @@ import java.util.Scanner;
 import com.cloud.chatbot.annotations.Nullable;
 import com.cloud.chatbot.exceptions.MissingFlagInputException;
 import com.cloud.chatbot.exceptions.MissingInputException;
-import com.cloud.chatbot.task.Deadline;
-import com.cloud.chatbot.task.Event;
-import com.cloud.chatbot.task.Item;
-import com.cloud.chatbot.task.ItemManager;
-import com.cloud.chatbot.task.Task;
+import com.cloud.chatbot.item.Deadline;
+import com.cloud.chatbot.item.Event;
+import com.cloud.chatbot.item.Item;
+import com.cloud.chatbot.item.ItemManager;
+import com.cloud.chatbot.item.Task;
 import com.cloud.chatbot.token.CommandManager;
 import com.cloud.chatbot.token.FlagManager;
 import com.cloud.chatbot.token.Token;
