+++ src/main/java/com/cloud/chatbot/item/ItemManager.java
@@ -1,4 +1,4 @@
-package com.cloud.chatbot.task;
+package com.cloud.chatbot.item;
 
 import java.util.ArrayList;
 import java.util.List;
